refactor: extract SearchDomainFilterPanel component from Perplexity model panel
